pyproject.toml
[tool.poetry]
name = "mochi_converter"
version = "0.1.0"
description = "Converts markdown files to mochi cards."
authors = ["Marcin Elantkowski <marcin.elantkowski@roche.com>"]

[tool.poetry.dependencies]
python = "^3.7"
edn_format = "^0.7.0"
requests = "^2.22.0"
black = "^19.10b0"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^5.1.4"
jupyter = "^1.0.0"

[tool.poetry.scripts]
convert = "mochi_converter.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
